feat(server): multi region support in dataloaders
