+++ packages/server/modules/auth/tests/integration/registration.spec.ts
@@ -165,7 +165,7 @@ describe('Server registration', () => {
           streamId: basicAdminStream.id
         },
         {
-          context: createTestContext({
+          context: await createTestContext({
             userId: newUser.id,
             auth: true,
             role: Roles.Server.User,
