+++ packages/server/modules/core/tests/favoriteStreams.spec.js
@@ -279,7 +279,7 @@ describe('Favorite streams', () => {
     before(async () => {
       apollo = {
         apollo: await buildApolloServer(),
-        context: createAuthedTestContext(me.id)
+        context: await createAuthedTestContext(me.id)
       }
 
       // Drop all favorites to ensure we don't favorite already favorited streams
@@ -448,7 +448,7 @@ describe('Favorite streams', () => {
           // "Log in" with other user
           const apollo = {
             apollo: await buildApolloServer(),
-            context: createAuthedTestContext(otherGuy.id)
+            context: await createAuthedTestContext(otherGuy.id)
           }
 
           const { data, errors } = await executeOperation(
@@ -474,7 +474,7 @@ describe('Favorite streams', () => {
     before(async () => {
       apollo = {
         apollo: await buildApolloServer(),
-        context: createTestContext()
+        context: await createTestContext()
       }
     })
